fix date bug

+++ core/retrofit/src/main/java/com/glebalekseevjk/core/retrofit/model/TodoElement.kt
@@ -7,9 +7,9 @@ data class TodoElement(
     @SerializedName("text") val text: String?,
     @SerializedName("importance") val importance: String?,
     @SerializedName("done") val done: Boolean?,
-    @SerializedName("deadline") val deadline: Int?,
+    @SerializedName("deadline") val deadline: Long?,
     @SerializedName("color") val color: String?,
     @SerializedName("last_updated_by") val lastUpdatedBy: String?,
-    @SerializedName("changed_at") val changedAt: Int?,
-    @SerializedName("created_at") val createdAt: Int?,
+    @SerializedName("changed_at") val changedAt: Long?,
+    @SerializedName("created_at") val createdAt: Long?,
 )
